requirements.txt
torchfile==0.1.0
matplotlib==2.2.3
torch==1.1.0
pandas==0.24.2
numpy==1.16.3
scikit_image==0.15.0
tqdm==4.26.0
easydict==1.9
seaborn==0.9.0
Pillow==7.2.0
heatmappy==0.2.6
skimage==0.0
tensorboardX==2.0
torchvision==0.6.1
PyYAML==5.3.1
